requirements.txt
googletrans==4.0.0rc1
joblib==1.1.0
nltk==3.6.5
pandas==1.3.4
psycopg2==2.9.3
python-decouple==3.8
scikit_learn==0.24.2

langdetect~=1.0.9
scikit-learn~=0.24.2